+++ blocks/form/src/index.tsx
@@ -10,6 +10,7 @@ import EnumInput from './components/EnumInput';
 import FileInput from './components/FileInput';
 import GeoCoordinatesInput from './components/GeoCoordinatesInput';
 import NumberInput from './components/NumberInput';
+import RadioInput from './components/RadioInput';
 import StringInput from './components/StringInput';
 import styles from './index.css';
 import ValidationError from './utils/ValidationError';
@@ -30,6 +31,7 @@ const inputs = {
   number: NumberInput,
   integer: NumberInput,
   boolean: BooleanInput,
+  radio: RadioInput,
 };
 
 const validateString: Validator = (field: StringField, event, value: string, remap) => {